requirements.txt
torch
torchvision
diffusers
Pillow
streamlit
transformers
